Ball types cleanup
