+++ crates/control/src/ball_filter.rs
@@ -4,7 +4,7 @@ use color_eyre::Result;
 use nalgebra::{Matrix2, Matrix4};
 use serde::{Deserialize, Serialize};
 
-use ball_filter::{BallFilter as BallFiltering, BallHypothesis, BallPosition};
+use ball_filter::{BallFilter as BallFiltering, BallHypothesis};
 use context_attribute::context;
 use coordinate_systems::{Ground, Pixel};
 use framework::{AdditionalOutput, HistoricInput, MainOutput, PerceptionInput};
@@ -12,8 +12,8 @@ use geometry::circle::Circle;
 use linear_algebra::{distance, IntoFramed, IntoTransform, Isometry2, Point2};
 use projection::{camera_matrices::CameraMatrices, camera_matrix::CameraMatrix, Projection};
 use types::{
-    ball::BallPercept,
-    ball_position::HypotheticalBallPosition,
+    ball_detection::BallPercept,
+    ball_position::{BallPosition, HypotheticalBallPosition},
     cycle_time::CycleTime,
     field_dimensions::FieldDimensions,
     limb::{is_above_limbs, Limb, ProjectedLimbs},
