+++ crates/control/src/penalty_shot_direction_estimation.rs
@@ -1,4 +1,3 @@
-use ball_filter::BallPosition;
 use color_eyre::Result;
 use context_attribute::context;
 use coordinate_systems::{Field, Ground};
@@ -7,6 +6,7 @@ use linear_algebra::{Isometry2, Point2};
 use serde::{Deserialize, Serialize};
 use spl_network_messages::{GamePhase, SubState, Team};
 use types::{
+    ball_position::BallPosition,
     field_dimensions::{FieldDimensions, Half},
     filtered_game_controller_state::FilteredGameControllerState,
     parameters::PenaltyShotDirectionParameters,
