+++ crates/control/src/role_assignment.rs
@@ -3,7 +3,6 @@ use std::{
     time::{Duration, SystemTime},
 };
 
-use ball_filter::BallPosition;
 use color_eyre::{eyre::WrapErr, Result};
 use serde::{Deserialize, Serialize};
 
@@ -17,6 +16,7 @@ use spl_network_messages::{
     SubState, Team,
 };
 use types::{
+    ball_position::BallPosition,
     cycle_time::CycleTime,
     fall_state::FallState,
     field_dimensions::FieldDimensions,
