+++ crates/hulk_behavior_simulator/src/server.rs
@@ -7,7 +7,6 @@ use crate::{
     cyclers::control::Database,
     robot::to_player_number,
     simulator::{Frame, Simulator},
-    state::Ball,
 };
 use color_eyre::{
     eyre::{Context, Error},
@@ -18,7 +17,10 @@ use path_serde::{PathDeserialize, PathIntrospect, PathSerialize};
 use serde::{Deserialize, Serialize};
 use tokio::{net::ToSocketAddrs, select, time::interval};
 use tokio_util::sync::CancellationToken;
-use types::{field_dimensions::FieldDimensions, hardware::Ids, players::Players};
+use types::{
+    ball_position::SimulatorBallState, field_dimensions::FieldDimensions, hardware::Ids,
+    players::Players,
+};
 
 #[derive(Clone, Serialize, Deserialize, PathSerialize, PathDeserialize, PathIntrospect)]
 struct Parameters {
