+++ crates/types/src/ball_detection.rs
