+++ crates/types/src/ball_position.rs
@@ -1,7 +1,27 @@
+use std::time::SystemTime;
+
+use coordinate_systems::Field;
 use serde::{Deserialize, Serialize};
 
-use linear_algebra::Point2;
+use linear_algebra::{Point2, Vector2};
 use path_serde::{PathDeserialize, PathIntrospect, PathSerialize};
+
+#[derive(
+    Debug, Clone, Copy, PathDeserialize, PathSerialize, PathIntrospect, Serialize, Deserialize,
+)]
+pub struct BallPosition<Frame> {
+    pub position: Point2<Frame>,
+    pub velocity: Vector2<Frame>,
+    pub last_seen: SystemTime,
+}
+
+#[derive(
+    Debug, Clone, Copy, PathDeserialize, PathSerialize, PathIntrospect, Serialize, Deserialize,
+)]
+pub struct SimulatorBallState {
+    pub position: Point2<Field>,
+    pub velocity: Vector2<Field>,
+}
 
 #[derive(
     Clone, Copy, Debug, Deserialize, Serialize, PathSerialize, PathDeserialize, PathIntrospect,
