+++ tools/twix/src/panels/map/layers/ball_measurements.rs
@@ -5,7 +5,7 @@ use eframe::epaint::{Color32, Stroke};
 
 use coordinate_systems::Ground;
 use linear_algebra::Point2;
-use types::{ball::BallPercept, field_dimensions::FieldDimensions};
+use types::{ball_detection::BallPercept, field_dimensions::FieldDimensions};
 
 use crate::{
     nao::Nao, panels::map::layer::Layer, twix_painter::TwixPainter, value_buffer::BufferHandle,
